refactor(development): split orchestrator into modules + remove stub agents
Break monolithic orchestrator.ts (2055 lines) into 5 focused modules:
- plan-snapshot.ts: plan validation and resolution helpers
- workspace.ts: sandbox management and file operations
- agent-executor.ts: agent task lifecycle (create/complete/execute)
- run-control.ts: public API (create/get/cancel/retry/recover/checkpoint)
- orchestrator.ts: core pipeline logic (~970 lines)

Remove 3 stub agents (ReleaseAgent, RecoveryAgent, DeployAgent) that
masked real logic. Release now calls gitops directly; deploy is inline;
recovery retry loop decides based on attempt count.
