+++ src/app/api/projects/[id]/development/runs/route.ts
@@ -2,7 +2,7 @@ import { auth } from '@clerk/nextjs/server'
 import { NextResponse } from 'next/server'
 import { prisma } from '@/lib/db/prisma'
 import { assertProjectOwnership } from '@/lib/development/auth'
-import { createDevelopmentRun, isDevelopmentRunActiveInWorker } from '@/lib/development/orchestrator'
+import { createDevelopmentRun, isDevelopmentRunActiveInWorker } from '@/lib/development/run-control'
 import { assertAutonomousDevelopmentSchemaReady } from '@/lib/development/schema-health'
 import type { ApprovedDevelopmentPlan } from '@/lib/development/types'
 import type { AssessmentResult, IterationPlanItem } from '@/types/development'
