+++ src/lib/development/agents.ts
@@ -559,51 +559,3 @@ export async function runReviewAgent(
   }
 }
 
-export async function runReleaseAgent(
-  _context: AgentExecutionContext,
-  iteration: IterationPlanItem,
-  branchName: string
-): Promise<AgentExecutionResult<Record<string, unknown>>> {
-  return {
-    output: {
-      branchName,
-      pullRequestTitle: `feat(iter-${iteration.index}): ${iteration.name}`,
-      mergeStrategy: 'squash',
-      merged: true,
-    },
-    tokenUsage: 90,
-    cost: 0.0007,
-  }
-}
-
-export async function runRecoveryAgent(
-  _context: AgentExecutionContext,
-  iteration: IterationPlanItem,
-  attempt: number,
-  failedGates: string[]
-): Promise<AgentExecutionResult<Record<string, unknown>>> {
-  return {
-    output: {
-      nextAttempt: attempt + 1,
-      recommendation: `Ajustar implementação da iteração ${iteration.index} para gates: ${failedGates.join(', ')}`,
-      fallback: 'checkpoint',
-    },
-    tokenUsage: 120,
-    cost: 0.0009,
-  }
-}
-
-export async function runDeployAgent(
-  _context: AgentExecutionContext,
-  iteration: IterationPlanItem
-): Promise<AgentExecutionResult<Record<string, unknown>>> {
-  return {
-    output: {
-      deployed: true,
-      message: `Deploy concluído para iteração ${iteration.index}`,
-      environment: 'production',
-    },
-    tokenUsage: 80,
-    cost: 0.0006,
-  }
-}
